Update dependencies, psalm, cs fixes

+++ module/Application/src/FileSize.php
@@ -18,7 +18,7 @@ class FileSize
 
         $bytes = max($bytes, 0);
         $pow   = floor(($bytes ? log($bytes) : 0) / log(1024));
-        $pow   = min($pow, count($units) - 1);
+        $pow   = (int) min($pow, count($units) - 1);
 
         $bytes /= pow(1024, $pow);
